third_party/WebKit/LayoutTests/fast/css/invalidation/empty-pseudo-sibling-expected.txt

Use descendant invalidation set for :empty pseudo class - sibling invalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(emptySibling, '').backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(emptySibling, '').backgroundColor is transparent
PASS getComputedStyle(notEmptySibling, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(notEmptySibling, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
